refactor and add console actions to one place
