+++ src/k8s/build.ts
@@ -191,16 +191,4 @@ export class Build {
         }
         return result;
     }
-
-    static async openConsole(context: { id: any; }) {
-        if (!context) {
-            vscode.window.showErrorMessage("Cannot load the Build");
-            return;
-        }
-        const k8sConfig = new KubeConfigUtils();
-        const clusterUrl = k8sConfig.getCurrentCluster().server;
-        const project = (k8sConfig.contexts).find((ctx) => ctx.name === k8sConfig.currentContext).namespace;
-        await open(`${clusterUrl}/console/project/${project}/browse/builds/${context.id}?tab=history`);
-        return;
-    }
 }
